refactor: Networkable 프로토콜 및 WeatherAPIProvider 생성
- NetworkManager가 Networkable 프로토콜 채택
