+++ Diary/Network/Networkable.swift
@@ -1,0 +1,37 @@
+//
+//  Networkable.swift
+//  Diary
+//
+//  Created by Aaron, Gundy, Rhovin on 2023/01/10.
+//
+
+import Foundation
+
+protocol Networkable { }
+
+extension Networkable {
+    func fetchData(url: URL?, completion: @escaping (Result<Data, Error>) -> Void) {
+        guard let url = url else {
+            completion(.failure(NetworkError.invalidURL))
+            return
+        }
+
+        let dataTask = URLSession.shared.dataTask(with: url) { data, response, error in
+            if let error = error {
+                completion(.failure(error))
+                return
+            }
+            guard let httpResponse = response as? HTTPURLResponse,
+                  (200...299).contains(httpResponse.statusCode) else {
+                completion(.failure(NetworkError.responseError))
+                return
+            }
+            guard let data = data else {
+                completion(.failure(NetworkError.invalidData))
+                return
+            }
+            completion(.success(data))
+        }
+        dataTask.resume()
+    }
+}
